+++ Kernel/platform-genie-eg64/main.c
@@ -160,9 +160,9 @@ int plt_rtc_write(void)
  *	the runtime
  */
 
-int strlen(const char *p)
+size_t strlen(const char *p)
 {
-  int len = 0;
+  size_t len = 0;
   while(*p++)
     len++;
   return len;
